buncha fixes and improvements

+++ shared/tokens.go
@@ -7,9 +7,9 @@ import (
 	"github.com/pkoukk/tiktoken-go"
 )
 
-const MaxTokens uint32 = 8000
+const MaxTokens int = 8000
 
-func GetNumTokens(text string) (numTokens uint32) {
+func GetNumTokens(text string) (numTokens int) {
 	tkm, err := tiktoken.EncodingForModel("gpt-4")
 	if err != nil {
 		err = fmt.Errorf("encoding for model: %v", err)
@@ -17,5 +17,5 @@ func GetNumTokens(text string) (numTokens uint32) {
 		return
 	}
 
-	return uint32(len(tkm.Encode(text, nil, nil)))
+	return len(tkm.Encode(text, nil, nil))
 }
